# lib/api-zod/src/generated/types/athleteUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AthleteInput } from './athleteInput';

export type AthleteUpdate = AthleteInput;